veris-engine/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
numpy
opencv-python-headless
imagehash
invisible-watermark
transformers
torch